Combine EKS and EKI to EnsembleKalmanProcess
