+++ src/CalibrateEmulateSample.jl
@@ -2,16 +2,15 @@ module CalibrateEmulateSample
 
 using Distributions, Statistics, LinearAlgebra, DocStringExtensions
 
-include("EKI.jl")
+include("EnsembleKalmanProcesses.jl")
+include("Observations.jl")
+include("Utilities.jl")
 include("GPEmulator.jl")
 include("MCMC.jl")
-include("Observations.jl")
 include("GModel.jl")
-include("Utilities.jl")
 include("spaces.jl")
 include("problems.jl")
 include("Histograms.jl")
-include("EKS.jl")
 include("EKS_bkp.jl")
 include("GPR.jl")
